# swagger.json
{
  "basePath": "/api",
  "host": "vulnerableapi.herokuapp.com",
  "consumes": [
    "application/json"
  ],
  "definitions": {
    "Pet": {
      "properties": {
        "breed": {
          "type": "string",
          "x-rs-sample-value":"Husky"
        },
        "id": {
          "format": "int64",
          "type": "integer",
          "x-rs-sample-value":-1
        },
        "name": {
          "type": "string"
        },
        "tag": {
          "type": "string"
        }
      },
      "required": [
        "id",
        "name"
      ],
      "type": "object",
      "xml": {
        "name": "Pet"
      }
    }
  },

  "info": {
    "contact": {
      "name": "API Team"
    },
    "description": "A sample API that uses a petstore as an example to demonstrate features in the swagger-2.0 specification",
    "license": {
      "name": "MIT"
    },
    "termsOfService": "",
    "title": "Swagger Petstore",
    "version": "1.0.0"
  },
  "paths": {
    "/pet": {
      "post": {
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "security": [
                    {

                    }
                ],
        "description": "",
        "operationId": "addPet",
        "parameters": [
          {
            "description": "Pet object that needs to be added to the store",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Pet"
            }
          }
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "summary": "Add a new pet to the store",
        "tags": [
          "pet post"
        ]
      }
    },
    "/pets": {
      "get": {
        "description": "Returns all pets from the system that the user has access to",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
                    {

                    }
                ],
        "responses": {
          "200": {
            "description": "A list of pets.",
            "schema": {
              "items": {
                "$ref": "#/definitions/Pet"
              },
              "type": "array"
            }
          }
        },
        "tags": [
          "pets get"
        ]
      }
    }
  },
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {

  },
  "schemes": [
    "https"
  ],


  "swagger": "2.0"
}